Fix the default status creation
The bug occours when launchsdi.sh is executed more than one time. When
first running the default status is changed to OFFLINE, but when running
again, without killing the connections, the OFFLINE state is forced but
the host can be ONLINE.

+++ launchsdi.sh
@@ -51,7 +51,9 @@ function createdatastructure()
             echo "<$FIELD value=\"\" />" > $DATAPATH/$FIELD.xml
         fi
     done
-    echo "<status value=\"OFFLINE\" class=\"red\" />" > $DATAPATH/status.xml
+    if ! test -f "$DATAPATH/status.xml"; then
+        echo "<status value=\"OFFLINE\" class=\"red\" />" > $DATAPATH/status.xml
+    fi
 }
 
 function getcolumns()
